+++ fractals/src/main/java/me/catmousedog/fractals/paneloperators/fractals/mandelbrot/Mandelbrot.java
@@ -12,7 +12,6 @@ import me.catmousedog.fractals.paneloperators.functions.IterativeFunction;
 import me.catmousedog.fractals.paneloperators.functions.LambertFunction;
 import me.catmousedog.fractals.paneloperators.functions.NormalizedFunction;
 import me.catmousedog.fractals.paneloperators.functions.PotentialFunction;
-import me.catmousedog.fractals.paneloperators.functions.TestFunction;
 
 public class Mandelbrot extends Fractal {
 
@@ -21,7 +20,7 @@ public class Mandelbrot extends Fractal {
 		items = null;
 		functions = new Function[] { new IterativeFunction(this), new NormalizedFunction(this),
 				new PotentialFunction(this), new EscapeAngleFunction(this), new BinaryFunction(this),
-				new LambertFunction(this), new DistanceEstimator(this), new TestFunction(this) };
+				new LambertFunction(this), new DistanceEstimator(this) };
 		function = functions[0];
 		mouse = null;
 	}